Add shadow JAR and missing plugins
Adding the shadow JAR plugin is technically violating the original
project requirements by adding an external dependency. However, it's
only being used to bundle all the code from each of the different
modules into a single archive so it can be run easily. So, I don't
think it's that big of a deal.